services/license_api/requirements.txt
alembic==1.16.5
fastapi==0.116.1
pydantic==2.11.7
pydantic-settings==2.10.1
psycopg[binary]==3.2.10
PyJWT==2.10.1
python-dotenv==1.1.1
sqlalchemy==2.0.43
uvicorn[standard]==0.35.0
